requirements.txt
colorama>=0.4.4
numpy>=1.19.2
opencv-python>=4.4.0.44
Pillow>=7.2.0
python-barcode>=0.13.1
pyzbar>=0.1.8
qrcode>=6.1
six>=1.15.0
